Move UInt implementation to a separate module
